backend/pyproject.toml
[project]
name = "tearsheet"
version = "0.1.0"
description = "Vietnam stock portfolio analysis web app using Flask, vnstock, and quantstats."
authors = [
    {name = "Your Name", email = "your@email.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.3",
    "vnstock>=3.2.6",
    "quantstats>=0.0.69",
    "pandas>=2.2.2",
    "numpy>=1.26.4,<2.0.0",
    "tenacity>=8.2.3",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.17.0",
    "kaleido>=0.2.1",
    "matplotlib>=3.10.1",
    "flask-cors>=6.0.1",
]

[dependency-groups]
dev = [
    # Changelog automation
    "gitchangelog>=3.0.4",
    
    # Code quality
    "ruff>=0.8.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    
    # Testing
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "pytest-flask>=1.3.0",
    "coverage>=7.0.0",
    
    # Documentation
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    
    # Development utilities
    "pre-commit>=3.0.0",
    "mypy>=1.8.0",
    "ipython>=8.24.0",
]


[tool.uv]
# Configure uv for optimal performance
concurrent-builds = 4
concurrent-downloads = 8
concurrent-installs = 4

# Cache configuration for better performance
cache-keys = [
    { file = "pyproject.toml" },
    { file = "requirements.txt" },
    { git = { commit = true } }
]
